+++ src/XUnity.AutoTranslator.Plugin.Core/Parsing/UnityTextParsers.cs
@@ -9,11 +9,13 @@ namespace XUnity.AutoTranslator.Plugin.Core.Parsing
    internal static class UnityTextParsers
    {
       public static RichTextParser RichTextParser;
+      public static RegexSplittingTextParser RegexSplittingTextParser;
       public static GameLogTextParser GameLogTextParser;
 
       public static void Initialize( Func<string, int, bool> isTranslatable )
       {
          RichTextParser = new RichTextParser();
+         RegexSplittingTextParser = new RegexSplittingTextParser();
          GameLogTextParser = new GameLogTextParser( isTranslatable );
       }
    }